Move input helpers to the inputs package
